refactor: CreatedEvent에서 View 부분 분리
