+++ client/src/components/Design/components/CreatedEventItem/CreatedEventItem.style.ts
